+++ xml/index.js
@@ -1,0 +1,41 @@
+import XMLAnchorNode from './XMLAnchorNode'
+import XMLAnnotationNode from './XMLAnnotationNode'
+import XMLContainerNode from './XMLContainerNode'
+import XMLDocument from './XMLDocument'
+import XMLElementNode from './XMLElementNode'
+import XMLElementNodeConverter from './XMLElementNodeConverter'
+import XMLExternalNode from './XMLExternalNode'
+import XMLExternalNodeConverter from './XMLExternalNodeConverter'
+import XMLSchema from './XMLSchema'
+import XMLTextNode from './XMLTextNode'
+import XMLTextNodeConverter from './XMLTextNodeConverter'
+import XMLValidator from './XMLValidator'
+import analyzeSchema from './analyzeSchema'
+import checkSchema from './checkSchema'
+import prettyPrint from './prettyPrint'
+import compileRNG from './compileRNG'
+import deserializeXMLSchema from './deserializeXMLSchema'
+import serializeXMLSchema from './serializeXMLSchema'
+import registerSchema from './registerSchema'
+
+export {
+  XMLAnchorNode,
+  XMLAnnotationNode,
+  XMLContainerNode,
+  XMLDocument,
+  XMLElementNode,
+  XMLElementNodeConverter,
+  XMLExternalNode,
+  XMLExternalNodeConverter,
+  XMLSchema,
+  XMLTextNode,
+  XMLTextNodeConverter,
+  XMLValidator,
+  analyzeSchema,
+  checkSchema,
+  prettyPrint as prettyPrintXML,
+  serializeXMLSchema,
+  deserializeXMLSchema,
+  compileRNG,
+  registerSchema
+}